Refactor deck layers config into multiple files
