+++ src/config/deck-layers/infrastructure-layer.ts
@@ -1,0 +1,13 @@
+import { MVTLayer } from 'deck.gl';
+
+export function infrastructureLayer(...props) {
+  return new MVTLayer(
+    {
+      binary: true,
+      autoHighlight: true,
+      highlightColor: [0, 255, 255, 255],
+      refinementStrategy: 'best-available',
+    } as any,
+    ...props,
+  );
+}
